separated add medicine dialog to separate file
